Add water space functionality
Adds functionality to plot ice diagram with relative humidity with respect to water instead of with respect to ice. Adds water RH info to structure created by makeGrowthDiagramStruct. Adds a "phaseFlag" input to growthDiagramProfile which switches between ice and water. iceGrowthDiagramWater makes a ice diagram figure with respect to water. This functionality may be rolled into iceGrowthDiagram in the long run.