Add BH04 lab results to makeGrowthDiagramStruct
Bailey and Hallett 2004's lab results extend to 200% ice saturations. Bailey and Hallett's 2009 ice diagram cuts off the diagram at 160%, and the code used to assume this was the maximum of known ice growth information. This push extends the supersaturations to the true maxima recorded in Bailey and Hallett 2004. These changes are also documented in the function help.

This doesn't make any huge changes to the science--these just extend the polycrystalline region to more extreme supersaturations.